add allow_all_drums_downbeat
This is kind of a hack, and doesn't properly understand all the different drum beats I can play. It's core problem is that if I play a drum on every server beat it will think down beats are coming much too often. But it does work for some patterns, and it may be good enough for now.